Fixed compilation issues
