+++ DCL_PiXYZ/PiXYZWorflow/PXZExporter.cs
@@ -17,7 +17,7 @@ namespace DCL_PiXYZ
         private readonly int lodLevel;
         private readonly SceneConversionPathHandler pathHandler;
 
-        public PXZExporter(PXZUtils pxzParams, SceneConversionPathHandler pathHandler, SceneConversionInfo sceneConversionInfo)
+        public PXZExporter(PXZConversionParams pxzParams, SceneConversionPathHandler pathHandler, SceneConversionInfo sceneConversionInfo)
         {
             this.pathHandler = pathHandler;
             path = pathHandler.OutputPath;
